{{ cookiecutter.repo_name }}/web/{{cookiecutter.repo_name}}/{{cookiecutter.repo_name}}/wsgi.py
import os
os.environ.setdefault('DJANGO_SETTINGS_MODULE',
                      '{{cookiecutter.repo_name}}.settings')

from configurations.wsgi import get_wsgi_application
application = get_wsgi_application()
