{{ cookiecutter.repo_name }}/web/requirements.txt
# Libraries
psycopg2==2.5.4
Unipath==1.0
hiredis==0.1.5

# Production
gunicorn==19.1.1
setproctitle==1.1.8
raven==5.1.1

# Development
ipython
ipdb
django-debug-toolbar==1.2.2

# Django
Django==1.7.1
django-configurations==0.8
dj-database-url==0.3.0
dj-email-url==0.0.2
django-redis-cache==0.13.0
